requirements.txt
flask
openai-whisper
googletrans==4.0.0rc1
monsterapi
python-dotenv
torch
ffmpeg-python